resolve code review issues
